feat: extract gateway_following_rest and mount via capability registry
