+++ packages/gateway_following_rest/pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "gateway_following_rest"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_rest_common", "gateway_following", "fastapi>=0.138.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+following_rest = "gateway_following_rest:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_following_rest"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_rest_common = { workspace = true }
+gateway_following = { workspace = true }
